bias_study/seat_test/requirements.txt
alabaster==0.7.12
allennlp==0.8.2
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==18.2.0
aws-xray-sdk==0.95
awscli==1.16.116
babel==2.6.0
boto==2.49.0
boto3==1.9.106
botocore==1.12.106
bz2file==0.98
chardet==3.0.4
click==7.0
colorama==0.3.9
conllu==0.11
cookies==2.2.1
cryptography==2.6.1
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
dill==0.2.9
docker==3.7.0
docker-pycreds==0.4.0
docutils==0.14
ecdsa==0.13
editdistance==0.5.2
flaky==3.5.3
flask==1.0.2
flask-cors==3.0.7
ftfy==5.5.1
future==0.17.1
gensim==3.7.1
gevent==1.3.6
greenlet==0.4.15
idna==2.8
imagesize==1.1.0
itsdangerous==1.1.0
jinja2==2.10
jmespath==0.9.4
jsondiff==1.1.1
jsonnet==0.10.0
jsonpickle==1.1
kiwisolver==1.0.1
markupsafe==1.1.1
matplotlib==2.2.3
mock==2.0.0
more-itertools==6.0.0
moto==1.3.4
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
nltk==3.4
numpydoc==0.8.0
overrides==1.9
packaging==19.0
parsimonious==0.8.0
pbr==5.1.3
plac==0.9.6
pluggy==0.9.0
preshed==2.0.1
py==1.8.0
pyaml==18.11.0
pyasn1==0.4.5
pycryptodome==3.7.3
pygments==2.3.1
pyparsing==2.3.1
pytest==4.3.0
python-dateutil==2.8.0
python-jose==2.0.2
pytorch-pretrained-bert==0.6.1
pytz==2017.3
pyyaml==3.13
regex
requests==2.21.0
responses==0.10.5
rsa==3.4.2
s3transfer==0.2.0
scikit-learn==0.20.3
singledispatch==3.4.0.3
smart-open==1.8.0
snowballstemmer==1.2.1
spacy==2.0.18
sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
sqlparse==0.2.4
tensorboardx==1.2
tensorflow-hub==0.3.0
thinc==6.12.1
toolz==0.9.0
tqdm==4.31.1
ujson==1.35
unidecode==1.0.23
urllib3==1.24.1
wcwidth==0.1.7
websocket-client==0.55.0
wrapt==1.10.11
xmltodict==0.12.0